modify some hook api to use in Android5.1

+++ app/src/main/java/com/android/reverse/collecter/DexFileInfo.java
@@ -3,16 +3,16 @@ package com.android.reverse.collecter;
 public class DexFileInfo {
 	
 	private String dexPath;
-	private int mCookie;
+	private long mCookie;
 	private ClassLoader defineClassLoader;
 
-	public DexFileInfo(String dexPath,int mCookie) {
+	public DexFileInfo(String dexPath, long mCookie) {
 		super();
 		this.dexPath = dexPath;
 		this.mCookie = mCookie;
 	}
 	
-	public DexFileInfo(String dexPath,int mCookie,ClassLoader classLoader) {
+	public DexFileInfo(String dexPath, long mCookie, ClassLoader classLoader) {
 		this(dexPath,mCookie);
 		this.defineClassLoader = classLoader;
 	}
@@ -21,7 +21,7 @@ public class DexFileInfo {
 		return dexPath;
 	}
 
-	public int getmCookie() {
+	public long getmCookie() {
 		return mCookie;
 	}
